feat: Dashboard の動的メンバー選択対応＆クライアント分離
- dashboard/page.tsx をサーバーコンポーネント化
- members_list から name を取得しオプション化
- DashboardClient コンポーネント新規作成（クライアント処理を分離）
- Dashboard の Storybook を DashboardClient 配下へ移動
- テストを DashboardClient 対象に更新
- 削除: 旧 index.stories.tsx
