+++ app/(feature)/dashboard/index.test.tsx
@@ -1,11 +1,17 @@
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import { default as Dashboard } from './page';
+import { DashboardClient } from './DashboardClient';
 
 jest.mock('next/navigation', () => jest.requireActual('next-router-mock'));
 
 describe('Dashboard', () => {
   const user = userEvent.setup();
+  const mockMemberOptions = [
+    { value: 'all', label: 'All' },
+    { value: 'eito', label: 'eito' },
+    { value: 'mei', label: 'mei' },
+    { value: 'tohhiro', label: 'tohhiro' },
+  ];
 
   afterEach(() => {
     jest.clearAllMocks();
@@ -13,13 +19,13 @@ describe('Dashboard', () => {
 
   describe('検索パネル', () => {
     test('SelectBoxが1つレンダリングされる', () => {
-      render(<Dashboard />);
+      render(<DashboardClient memberOptions={mockMemberOptions} />);
       const select = screen.getAllByRole('combobox');
       expect(select).toHaveLength(1);
     });
 
     test('Inputがdate属性で2つレンダリングされる', () => {
-      render(<Dashboard />);
+      render(<DashboardClient memberOptions={mockMemberOptions} />);
       const startInput = screen.getByLabelText('開始');
       const endInput = screen.getByLabelText('終了');
       [startInput, endInput].forEach((input) => {
@@ -29,7 +35,7 @@ describe('Dashboard', () => {
     });
 
     test('Buttonがsubmit属性、Enabledで1つレンダリングされる', () => {
-      render(<Dashboard />);
+      render(<DashboardClient memberOptions={mockMemberOptions} />);
       const button = screen.getByRole('button', { name: '検索' });
       expect(button).toBeInTheDocument();
       expect(button).toHaveAttribute('type', 'submit');
@@ -37,7 +43,7 @@ describe('Dashboard', () => {
     });
 
     test('対象者を未選択、開始終了をクリアし検索ボタンを押すとvalidationエラーになる', async () => {
-      render(<Dashboard />);
+      render(<DashboardClient memberOptions={mockMemberOptions} />);
       const validationsText = ['対象を選択', '開始日を選択', '終了日を選択'];
       const button = screen.getByRole('button', { name: '検索' });
 
@@ -54,14 +60,14 @@ describe('Dashboard', () => {
 
     describe('非表示/表示ボタン', () => {
       test('非表示のボタンがレンダリングされる', () => {
-        render(<Dashboard />);
+        render(<DashboardClient memberOptions={mockMemberOptions} />);
         const button = screen.getByRole('button', { name: '表示' });
         expect(button).toBeInTheDocument();
       });
 
       test('表示ボタンをクリックすると検索パネルが表示される', async () => {
         const labelNames = ['対象者を選択', '開始', '終了', '検索', '非表示'];
-        render(<Dashboard />);
+        render(<DashboardClient memberOptions={mockMemberOptions} />);
         const button = screen.getByRole('button', { name: '表示' });
 
         await user.click(button);
